requirements.txt
coreapi==2.3.3
Django==4.1.3
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-ipware==4.0.0
djangorestframework==3.13.1
gunicorn==19.6.0
psycopg2==2.9.3
pyyaml==6.0
requests==2.26.0
python-json-logger==2.0.2
tblib==1.7.0
APScheduler==3.10.1
firebase-admin==6.2.0